cleaned up st-wl.c
